# alafia-server/src/main/java/es/alafia/server/model/DinnerTable.java
package es.alafia.server.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.springframework.data.annotation.Id;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class DinnerTable {

    @Id
    private String id;

    private Booking booking;
}