# alafia-server/src/main/java/es/alafia/server/model/Drink.java
package es.alafia.server.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.springframework.data.annotation.Id;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class Drink {

    @Id
    private String id;

    private String description;

    private Double price;
}